requirements.txt
streamlit
streamlit-drawable-canvas
Pillow
numpy
tensorflow